Split log.h to speed up compilation (#15258)
