+++ src/client/sound/ogg_file.cpp
@@ -26,6 +26,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #include <cassert>
 #include <cstring> // memcpy
+#include <memory>
 
 namespace sound {
