+++ src/client/sound/sound_singleton.h
@@ -24,6 +24,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #pragma once
 
+#include <memory>
 #include "al_helpers.h"
 
 namespace sound {
